Grr, lots more fixes:
- Leftover ORG_SLAMB_ stuff.
- Some missing underscores.
- The ".comm" sections apparently weren't necessary at all - it must only be
  Darwin that cares about defined symbols in assembly. And by link time,
  they're there.
- Fixed a SIGSEGV if the TSD is not set at time of exit in single-threaded
  builds.
- Alpha single-threaded fix; it was derefencing where it shouldn't.

I'd say this is the last round of fixes from that change, but it's not. The
alpha code isn't working now. Not sure why.

+++ tests/platform_behavior/test_socket_timeout.c
@@ -43,7 +43,7 @@ double doublify_timeval(const struct timeval *tv) {
 
 int main(void) {
     struct sockaddr_in server_addr;
-    int server_addr_len = sizeof(struct sockaddr_in);
+    unsigned long server_addr_len = sizeof(struct sockaddr_in);
     int listen_fd, connected_fd;
     int retval;
     struct timeval timeout, old_time, new_time;
@@ -60,7 +60,7 @@ int main(void) {
     error_wrap(listen(listen_fd, 5), "listen");
     if (error_wrap(fork(), "fork") == 0) {
         struct sockaddr_in originating_addr;
-        int originating_addr_len = sizeof(struct sockaddr_in);
+        unsigned long originating_addr_len = sizeof(struct sockaddr_in);
         int accepted_fd;
 
         /* Accept the connection, wait for it to be closed, then die. */
